Add an example of calling the backend for sampling
