+++ src/SpecComponent.tsx
@@ -9,7 +9,7 @@ const math11 = require('mathjs11');
 
 function SpecComponent() {
   const { spec: value, setSpec: setValue } = useContext(SpecContext);
-  const [spec, setSpec] = useState(value);
+  const [spec, setSpec] = useState(value || new Spec('sqrt(x + 1) - sqrt(x)', [new SpecRange('x', -1e308, 1e308, 0)], 0));
   
   // When the spec is clicked, we show an overlay menu for editing the spec and the input ranges for each variable.
   const [showOverlay, setShowOverlay] = useState(false);
@@ -19,6 +19,10 @@ function SpecComponent() {
   }
   const handleOverlayClick = () => {
     setShowOverlay(false);
+  }
+  // Wait until submit click to set the spec
+  const handleSubmitClick = () => {
+    setValue(spec);
   }
   
   function getVariables(spec: Spec) : string[] {
@@ -41,7 +45,7 @@ function SpecComponent() {
         <div className="spec-overlay-content" onClick={(event) => event.stopPropagation()}>
           <div className="spec-overlay-header">
             <div>Spec</div>
-            <button onClick={handleOverlayClick}>Done</button>
+            <button onClick={handleSubmitClick}>Done</button>
           </div>
           {/* Render the expression into HTML with KaTeX */}
           <div className="spec-tex" dangerouslySetInnerHTML={{ __html: (() => { try { return KaTeX.renderToString(math11.parse(spec.expression).toTex(), { throwOnError: false }) } catch (e) { return (e as Error).toString() } })() }} />
